Move transformation-layers to sub package (#710)
* First pass over moving transforms to separate package

* Fixing wrong imports

* Updating imports in sample_batch

* Fixing failing tests

* Trying to fix failing example tests

* Trying to fix failing example tests

* Adding copyright

* Rename ToDense, ToSparse & ToRagged, to make it clear it's for list-columns
