+++ merlin/models/tf/inputs/embedding.py
@@ -35,7 +35,7 @@ from merlin.models.tf.core.tabular import (
     TabularAggregationType,
     TabularBlock,
 )
-from merlin.models.tf.core.transformations import AsDenseFeatures, AsSparseFeatures
+from merlin.models.tf.transforms.tensor import ListToDense, ListToSparse
 
 # pylint has issues with TF array ops, so disable checks until fixed:
 # https://github.com/PyCQA/pylint/issues/3613
@@ -647,7 +647,7 @@ class EmbeddingFeatures(TabularBlock):
         **kwargs,
     ):
         if add_default_pre:
-            embedding_pre = [Filter(list(feature_config.keys())), AsSparseFeatures()]
+            embedding_pre = [Filter(list(feature_config.keys())), ListToSparse()]
             pre = [embedding_pre, pre] if pre else embedding_pre  # type: ignore
         self.feature_config = feature_config
         self.l2_reg = l2_reg
@@ -1006,7 +1006,7 @@ class SequenceEmbeddingFeatures(EmbeddingFeatures):
         **kwargs,
     ):
         if add_default_pre:
-            embedding_pre = [Filter(list(feature_config.keys())), AsDenseFeatures(max_seq_length)]
+            embedding_pre = [Filter(list(feature_config.keys())), ListToDense(max_seq_length)]
             pre = [embedding_pre, pre] if pre else embedding_pre  # type: ignore
 
         super().__init__(
