+++ merlin/models/tf/prediction_tasks/next_item.py
@@ -22,21 +22,59 @@ from tensorflow.keras.layers import Layer
 from merlin.models.tf.blocks.retrieval.base import ItemRetrievalScorer
 from merlin.models.tf.blocks.sampling.cross_batch import PopularityBasedSampler
 from merlin.models.tf.core.base import Block
-from merlin.models.tf.core.transformations import (
-    ItemsPredictionWeightTying,
-    L2Norm,
-    LabelToOneHot,
-    LogitsTemperatureScaler,
-    PopularityLogitsCorrection,
-)
 from merlin.models.tf.prediction_tasks.classification import (
     CategFeaturePrediction,
     MultiClassClassificationTask,
 )
+from merlin.models.tf.transforms.bias import LogitsTemperatureScaler, PopularityLogitsCorrection
+from merlin.models.tf.transforms.features import ToOneHot
+from merlin.models.tf.transforms.regularization import L2Norm
+from merlin.models.utils import schema_utils
 from merlin.models.utils.schema_utils import categorical_cardinalities, categorical_domains
 from merlin.schema import Schema, Tags
 
 LOG = logging.getLogger("merlin.models")
+
+
+@Block.registry.register_with_multiple_names("weight-tying")
+@tf.keras.utils.register_keras_serializable(package="merlin_models")
+class ItemsPredictionWeightTying(Block):
+    """Tying the item embedding weights with the output projection layer matrix [1]
+    The output logits are obtained by multiplying the output vector by the item-ids embeddings.
+    Parameters
+    ----------
+        schema : Schema
+            The `Schema` with the input features
+        bias_initializer : str, optional
+            Initializer to use on the bias vector, by default "zeros"
+    References:
+    -----------
+    [1] Hakan, Inan et al.
+        "Tying word vectors and word classifiers: A loss framework for language modeling"
+        arXiv:1611.01462
+    """
+
+    def __init__(self, schema: Schema, bias_initializer="zeros", **kwargs):
+        super(ItemsPredictionWeightTying, self).__init__(**kwargs)
+        self.bias_initializer = bias_initializer
+        self.item_id_feature_name = schema.select_by_tag(Tags.ITEM_ID).column_names[0]
+        self.num_classes = schema_utils.categorical_cardinalities(schema)[self.item_id_feature_name]
+        self.item_domain = schema_utils.categorical_domains(schema)[self.item_id_feature_name]
+
+    def build(self, input_shape):
+        self.bias = self.add_weight(
+            name="output_layer_bias",
+            shape=(self.num_classes,),
+            initializer=self.bias_initializer,
+        )
+        return super().build(input_shape)
+
+    def call(self, inputs, training=False, **kwargs) -> tf.Tensor:
+        embedding_table = self.context.get_embedding(self.item_domain)
+        logits = tf.matmul(inputs, embedding_table, transpose_b=True)
+        logits = tf.nn.bias_add(logits, self.bias)
+
+        return logits
 
 
 @tf.keras.utils.register_keras_serializable(package="merlin_models")
@@ -63,7 +101,6 @@ def ItemsPredictionPopSampled(
     That implementation does not require the actual item frequencies/probabilities
     if the item ids are sorted by frequency. The PopularityBasedSampler
     approximates the item probabilities using the log_uniform (zipfian) distribution.
-
     Parameters:
     -----------
         schema: Schema
@@ -77,22 +114,18 @@ def ItemsPredictionPopSampled(
         ignore_false_negatives: bool
             Ignore sampled items that are equal to the target classes
             Defaults to True
-
     Returns:
     -------
         A SequenceBlock that performs popularity-based sampling of negatives, scores
         the items and applies the logQ correction for sampled softmax
-
     References
     ----------
     .. [1] Yoshua Bengio and Jean-Sébastien Sénécal. 2003. Quick Training of Probabilistic
        Neural Nets by Importance Sampling. In Proceedings of the conference on Artificial
        Intelligence and Statistics (AISTATS).
-
     .. [2 Y. Bengio and J. S. Senecal. 2008. Adaptive Importance Sampling to Accelerate
        Training of a Neural Probabilistic Language Model. Trans. Neur. Netw. 19, 4 (April
        2008), 713–722. https://doi.org/10.1109/TNN.2007.912312
-
     .. [3] Jean, Sébastien, et al. "On using very large target vocabulary for neural
         machine translation." arXiv preprint arXiv:1412.2007 (2014).
     """
@@ -187,13 +220,9 @@ def NextItemPredictionTask(
         )
 
     else:
-        if weight_tying:
-            prediction_call = ItemsPredictionWeightTying(schema)
+        prediction_call = ItemsPrediction(schema)
 
-        else:
-            prediction_call = ItemsPrediction(schema)
-
-        prediction_call = prediction_call.connect(LabelToOneHot())
+        prediction_call = prediction_call.connect(ToOneHot())
 
     if post_logits is not None:
         prediction_call = prediction_call.connect(post_logits)
