+++ merlin/models/tf/transforms/negative_sampling.py
@@ -25,7 +25,7 @@ from merlin.schema import Schema, Tags
 
 
 @tf.keras.utils.register_keras_serializable(package="merlin.models")
-class UniformNegativeSampling(tf.keras.layers.Layer):
+class InBatchNegatives(tf.keras.layers.Layer):
     """Random in-batch negative sampling.
 
     Only works with positive-only binary-target batches.
@@ -56,7 +56,7 @@ class UniformNegativeSampling(tf.keras.layers.Layer):
         **kwargs
     ):
 
-        super(UniformNegativeSampling, self).__init__(**kwargs)
+        super(InBatchNegatives, self).__init__(**kwargs)
         self.n_per_positive = n_per_positive
         self.item_id_col = schema.select_by_tag(Tags.ITEM_ID).column_names[0]
         self.schema = schema.select_by_tag(Tags.ITEM)
