+++ merlin/models/tf/transforms/noise.py
@@ -13,7 +13,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
 from typing import Optional
 
 import tensorflow as tf
