+++ merlin/models/tf/transforms/regularization.py
@@ -1,0 +1,42 @@
+#
+# Copyright (c) 2021, NVIDIA CORPORATION.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+from typing import Union
+
+import tensorflow as tf
+
+from merlin.models.tf.core.base import Block
+from merlin.models.tf.core.combinators import TabularBlock
+from merlin.models.tf.typing import TabularData
+
+
+@Block.registry.register_with_multiple_names("l2-norm")
+@tf.keras.utils.register_keras_serializable(package="merlin.models")
+class L2Norm(TabularBlock):
+    """Apply L2-normalization to input tensors along a given axis"""
+
+    def __init__(self, **kwargs):
+        super(L2Norm, self).__init__(**kwargs)
+
+    def call(self, inputs: Union[tf.Tensor, TabularData], axis: int = -1, **kwargs):
+        if isinstance(inputs, dict):
+            inputs = {key: tf.linalg.l2_normalize(inp, axis=axis) for key, inp in inputs.items()}
+        else:
+            inputs = tf.linalg.l2_normalize(inputs, axis=axis)
+
+        return inputs
+
+    def compute_output_shape(self, input_shape):
+        return input_shape
